Bugfix: path params were being dropped.
request.params memoizes the union of GET and POST, and causes the router
to not inject path params into the hash. Instead, we now check only
POST params. Checking GET requests also seems a bit redundant since it's
not supposed to be used for the kind of requests BotAway is meant to
intercept.

Switched to BotAway::Middleware instead of overriding
ActionDispatch::ParamsParser. This is the "right way" to do it and is
also a much more robust solution.

This closes #2.